rename .d.ts => .ts
